requirements.txt
alpaca-py
pandas
numpy
yfinance
python-dotenv
requests
streamlit
altair
